+++ spec/hcloud/concerns/updatable_spec.rb
@@ -19,5 +19,15 @@ RSpec.describe HCloud::Updatable do
 
       expect(resource.name).to eq "my_name"
     end
+
+    it "updates the resource with the given attributes" do
+      stub_request(:put, "https://api.hetzner.cloud/v1/examples/#{resource.id}")
+        .with(body: resource.attributes.slice(*resource.updatable_attributes.map(&:to_s)).merge("name" => "MY_NAME"))
+        .to_return(body: { example: resource.attributes.merge(name: "my_name") }.to_json)
+
+      resource.update(name: "MY_NAME")
+
+      expect(resource.name).to eq "my_name"
+    end
   end
 end